:fire: removed old code for the charts configuration in the backend
