+++ server/charts/AxisChart.js
@@ -1,8 +1,8 @@
 const _ = require("lodash");
 const moment = require("moment");
-const NewBarChart = require("./NewBarChart");
-const NewLineChart = require("./NewLineChart");
-const NewPieChart = require("./NewPieChart");
+const BarChart = require("./BarChart");
+const LineChart = require("./LineChart");
+const PieChart = require("./PieChart");
 const determineType = require("../modules/determineType");
 const dataFilter = require("./dataFilter");
 
@@ -189,13 +189,13 @@ class AxisChart {
     let chart;
     switch (this.chart.type) {
       case "line":
-        chart = new NewLineChart(this.chart, this.datasets, this.axisData);
+        chart = new LineChart(this.chart, this.datasets, this.axisData);
         break;
       case "bar":
-        chart = new NewBarChart(this.chart, this.datasets, this.axisData);
+        chart = new BarChart(this.chart, this.datasets, this.axisData);
         break;
       default:
-        chart = new NewPieChart(this.chart, this.datasets, this.axisData);
+        chart = new PieChart(this.chart, this.datasets, this.axisData);
         break;
     }
